+++ ICD.Connect.Audio.Biamp/BiampTesiraDevice.cs
@@ -74,6 +74,7 @@ namespace ICD.Connect.Audio.Biamp
 
 		// Used with settings
 		private string m_Config;
+		private readonly IcdHashSet<IDeviceControl> m_LoadedControls;
 
 		#region Properties
 
@@ -136,6 +137,10 @@ namespace ICD.Connect.Audio.Biamp
 		/// </summary>
 		public BiampTesiraDevice()
 		{
+			m_LoadedControls = new IcdHashSet<IDeviceControl>();
+
+			Controls.Add(new BiampTesiraRoutingControl(this, 0));
+
 			m_SubscriptionCallbacks = new Dictionary<string, IcdHashSet<SubscriptionCallbackInfo>>();
 			m_SubscriptionCallbacksSection = new SafeCriticalSection();
 
@@ -282,10 +287,14 @@ namespace ICD.Connect.Audio.Biamp
 		/// <param name="xml"></param>
 		private void ParseXml(string xml)
 		{
-			Controls.Clear();
+			DisposeLoadedControls();
 
+			// Load and add the new controls
 			foreach (IDeviceControl control in ControlsXmlUtils.GetControlsFromXml(xml, m_AttributeInterfaces))
+			{
 				Controls.Add(control);
+				m_LoadedControls.Add(control);
+			}
 		}
 
 		#endregion
@@ -461,6 +470,17 @@ namespace ICD.Connect.Audio.Biamp
 		private void Initialize()
 		{
 			Initialized = true;
+		}
+
+		private void DisposeLoadedControls()
+		{
+			// Remove the previously loaded controls
+			foreach (IDeviceControl control in m_LoadedControls)
+			{
+				Controls.Remove(control.Id);
+				control.Dispose();
+			}
+			m_LoadedControls.Clear();
 		}
 
 		/// <summary>
@@ -739,6 +759,7 @@ namespace ICD.Connect.Audio.Biamp
 			base.ClearSettingsFinal();
 
 			m_Config = null;
+			DisposeLoadedControls();
 			Username = null;
 			SetPort(null);
 		}